skill_recommend_server/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "skills-recommender"
version = "1.0.0"
description = "Agent Skills 推荐系统"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "PyYAML>=6.0",
    "python-dotenv>=1.0.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",
    "FlagEmbedding>=1.2.10",
    "openai>=1.10.0",
    "anthropic>=0.18.0",
    "requests>=2.31.0",
    "langfuse>=2.0.0",
]

[project.scripts]
skills-recommender = "skills_recommender:main"

[tool.setuptools.packages.find]
include = ["skills_recommender*"]
